Rename texrenderer
