+++ src/qutip_qip/circuit/tex_renderer.py
